change of images, database laid out, and seeds made but couldn't run

+++ migrations/20160308160830_create_products.js
@@ -2,7 +2,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('products', function(table){
     table.increments('id');
     table.string('name');
-    table.decimal('price');
+    table.integer('price');
     table.string('image');
     table.text('description');
   });
